Fix bug that makes the game continue after a 5 win
There is no reason to play again after we have player gets to 5 point.
We need to restart the game.
So i made a return in the event handler - which is a parent to the
buttons. we check if any of the scores is 5 and retun the message to
dom. read code to understand more